Move init.lua content to main.lua, making everything reloadable
